Adds new N generic parameter to GossipSimulator
Enables defining the exact type of nodes
the simulator will deal with.
This way, when calling methods such as
simulator.getNodes(), the exact type
of nodes is returned.
If a new subclass of nodes is created.
It will be possible to access
their specific methods,
even if a node is accessed through the
simulator instance.

Signed-off-by: Manoel Campos <manoelcampos@gmail.com>